Rename directory audit_trail to audittrail
