+++ main.go
@@ -5,7 +5,7 @@ import (
 	"os"
 	"time"
 
-	audittrail "github.com/spacelift-io/spacectl/internal/cmd/audit_trail"
+	"github.com/spacelift-io/spacectl/internal/cmd/audittrail"
 	"github.com/spacelift-io/spacectl/internal/cmd/blueprint"
 	"github.com/spacelift-io/spacectl/internal/cmd/completion"
 	"github.com/spacelift-io/spacectl/internal/cmd/module"
